pyproject.toml
[project]
name = "goosebot"
version = "0.1.0"
description = "Discord bot that bridges Discord and Goose.ai CLI"
readme = "README.md"
requires-python = ">=3.11"
 dependencies = [
     "discord.py>=2.3.0",
     "python-dotenv>=1.0.0",
 ]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = []
